requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.2.0
tqdm>=4.65.0

# Analysis and visualization
umap-learn>=0.5.0

# Compendium generation
jinja2>=3.1.0
pyyaml>=6.0

# Dashboard (optional)
flask>=2.3.0
streamlit>=1.25.0

# Development
jupyter>=1.0.0
ipython>=8.0.0
